requirements.txt
openai==1.35.3
python-dotenv==1.0.0